requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3
geopandas==0.14.1
rasterio==1.3.9
shapely==2.0.2
folium==0.15.1

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
xgboost==2.0.3

# Databricks integration
databricks-sql-connector==3.0.0
databricks-connect==13.3.0

# Data sources APIs
requests==2.31.0
aiohttp==3.9.1
sentinelhub==3.10.0
earthengine-api==0.1.370

# Geospatial and mapping
pyproj==3.6.1
fiona==1.9.5
cartopy==0.22.0
matplotlib==3.8.2
seaborn==0.13.0

# Database and storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1